# openapi.yaml
openapi: 3.0.3
info:
  title: Template Builder API
  version: 1.0.0
  description: >
    Intent-driven, AI-native template builder and document generation API.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/templates/{templateId}/tests:
    post:
      summary: Create a test case for a template
      parameters:
        - in: path
          name: templateId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateTestCreateRequest"
      responses:
        "201":
          description: Test created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateTest"
...
components:
  schemas:
...
    TemplateTestCreateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        runtime_params:
          type: object
        expected_checks:
          type: object
    TemplateTest:
      type: object
      properties:
        test_id:
          type: string
          format: uuid
        template_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        runtime_params:
          type: object
        expected_checks:
          type: object
...